+++ samples/WeihanLi.EntityFramework.Core3_0Sample/TestDbContext.cs
@@ -2,30 +2,31 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using Microsoft.EntityFrameworkCore;
+using WeihanLi.EntityFramework.Audit;
 
-namespace WeihanLi.EntityFramework.Samples
+// ReSharper disable once CheckNamespace
+namespace WeihanLi.EntityFramework
 {
-    public class TestDbContext : DbContext
+    public class TestDbContext : AuditDbContext
     {
         public TestDbContext(DbContextOptions<TestDbContext> options) : base(options)
         {
         }
 
         public DbSet<TestEntity> TestEntities { get; set; }
-
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-        }
     }
 
+    [Table("tabTestEntities")]
     public class TestEntity
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
+        [Column("PKID")]
         public int Id { get; set; }
 
+        [Column("ExtraSettings")]
         public string Extra { get; set; }
 
-        public DateTime CreatedAt { get; set; }
+        public DateTimeOffset CreatedAt { get; set; }
     }
 }