+++ src/WeihanLi.EntityFramework/IEFUnitOfWork.cs
@@ -7,5 +7,7 @@ namespace WeihanLi.EntityFramework
     public interface IEFUnitOfWork<out TDbContext> : IUnitOfWork, IDisposable where TDbContext : DbContext
     {
         TDbContext DbContext { get; }
+
+        IEFRepository<TDbContext, TEntity> GetRepository<TEntity>() where TEntity : class;
     }
 }